% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyHSD.alphabet.R
\name{TukeyHSD_alphabet}
\alias{TukeyHSD_alphabet}
\title{Make groups for the results of TukeyHSD() with alphabets}
\usage{
TukeyHSD_alphabet(aov_res, factor_name, alpha = 0.05, ...)
}
\arguments{
\item{aov_res}{An object created with aov().}

\item{factor_name}{The name of explanatory variable you want to test.}

\item{alpha}{The significance level in multi-comparison test.}

\item{...}{Other parameters for TukeyHSD().}
}
\description{
Make groups for the results of TukeyHSD() with alphabets
}
